backend/pyproject.toml
[project]
name = "askdocs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "docx2txt>=0.9",
    "dynaconf>=3.2.11",
    "fastapi>=0.115.12",
    "fastembed>=0.4.2",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.76",
    "langchain-google-genai>=2.1.4",
    "langchain-huggingface>=0.2.0",
    "langchain-openai>=0.3.33",
    "langchain-qdrant>=0.2.1",
    "langchain-redis>=0.2.4",
    "langchain-unstructured>=0.1.6",
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pymupdf>=1.26.4",
    "pypdf>=5.5.0",
    "pytesseract>=0.3.10",
    "python-docx>=1.1.2",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "python-pptx>=1.0.2",
    "qdrant-client>=1.15.1",
    "redis>=6.4.0",
    "rq>=2.4.1",
    "sqlmodel>=0.0.24",
    "streamlit>=1.45.1",
    "unstructured>=0.17.2",
    "uvicorn>=0.34.2",
]
